% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basecost}
\alias{basecost}
\title{Empirical fire management costs.}
\format{A data frame with 7 rows of annual cost observations for each of four columns of FMO zones.}
\usage{
basecost
}
\description{
A data frame of empirical historical fire management costs in dollars per acre.
}
\keyword{datasets}
