% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdba}
\alias{cdba}
\title{Coniferous:deciduous burn area.}
\format{A data frame with 1,160 rows and 5 columns.}
\usage{
cdba
}
\description{
A data frame of annual Alaska coniferous and deciduous burn area total computed from ALFRESCO model output, by ALFRESCO FMO treatment and RCP.
}
\keyword{datasets}
