% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdratio}
\alias{cdratio}
\title{Coniferous:deciduous ratios.}
\format{A data frame with 580 rows and 4 columns.}
\usage{
cdratio
}
\description{
A data frame of annual Alaska coniferous:deciduous landscape cover area ratios computed from ALFRESCO model output, by ALFRESCO FMO treatment and RCP.
}
\keyword{datasets}
