% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cost}
\alias{cost}
\title{Historical and projected estimated fire management costs.}
\format{A data frame with 1,740 rows and 5 columns pertaining to ALFRESCO simulation FMO treatment, RCP, year, statistic and value.}
\usage{
cost
}
\description{
A data frame of estimated fire management costs in millions of dollars based on MC sampling of empirical costs and ALFRESCO simulated landscape fire scar model output.
}
\keyword{datasets}
