% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{costSummary}
\alias{costSummary}
\title{Projected estimated fire management costs.}
\format{A data frame with 48 rows of decadal averages of annual cost statistics pertaining to ALFRESCO simulation set, FMO treatment, RCP, and decade.}
\usage{
costSummary
}
\description{
A data frame of estimated fire management costs in millions of dollars based on MC sampling of empirical costs and ALFRESCO simulated landscape fire scar model output.
Decadal averages beginning with the 2020s of annual mean, 5th and 95th percentile costs are provided in this simplified summary of the \code{cost} data set.
}
\keyword{datasets}
