% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\name{datadoc}
\alias{datadoc}
\title{Generate a report on a data set}
\usage{
datadoc(data, file, title = "", subtitle = "", metadata = FALSE,
  open_doc = TRUE, keep_rmd = FALSE, details = TRUE)
}
\arguments{
\item{data}{jfsp data set, e.g., \code{fmoba}.}

\item{file}{character, output file, must end in .pdf, .docx, or .html. See details.}

\item{title}{character, document title.}

\item{subtitle}{character, document subtitle. Set to \code{NULL} to recover the \code{dataMaid} attribution.}

\item{metadata}{logical, if \code{TRUE}, include metadata. See details.}

\item{open_doc}{logical, open document after rendering. Defaults to \code{TRUE}.}

\item{keep_rmd}{logical, retain the intermediary Rmd document. Defaults to \code{FALSE}.}

\item{details}{logical, if set to \code{FALSE}, do not print any progress.}
}
\value{
nothing is returned, but a file is written to disk.
}
\description{
Generate a pdf, Word or HTML report on any data set in the jfsp package.
}
\details{
Create a report for a specific data set. The report summarizes the variables in the data set. It is a wrapper around and based on \code{dataMaid::makeDataReport}, though it renders output with some differences not available in \code{dataMaid}.

Pandoc must be available on the system for rendering to pdf. It is recommended to run from RStudio, which includes the necessary components, if you have difficulty.
For Word documents (docx) it is necessary to be on a Windows system with Microsoft Word installed.
If no file name is supplied, the output defaults to HTML and the file name is the name of the data set.

If \code{metadata = TRUE}, information about the system environment and conditions at run time when the document was generated is included at the bottom of the document. If \code{FALSE} (default), they are removed.
Because \code{datadoc} does not directly render the document with \code{dataMaid} and makes some internal changes to the Rmd file, the function call line in the metadata is always removed due to its resulting lack of clarity.
}
\examples{
\dontrun{datadoc(fmoba)}
}
