% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbxfire}
\alias{fbxfire}
\title{Probability of fire near Fairbanks, Alaska.}
\format{A data frame with 1,900 rows and 5 columns.}
\usage{
fbxfire
}
\description{
A data frame of the probability of fire near Fairbanks, Alaska within each of a sequence of radial buffers of Fairbanks, based on ALFRESCO simulation output.
Probabilities as a function of radius are broken out by FMO treatment, RCP, and different periods of time.
}
\keyword{datasets}
