% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{firesize}
\alias{firesize}
\title{Alaska fire sizes.}
\format{A data frame with 63,200 rows and 5 columns.}
\usage{
firesize
}
\description{
A data frame of the decadal distribution of fire sizes based on ALFRESCO simulation output, by FMO treatment and RCP.
The FS column gives the fire size in acres and the Freq column gives the frequency of fires that size over a decade and multiple ALFRESCO simulation replicates.
Frequencies should be treated as simple weights and can be rescaled if desired. Freq is an average frequency across GCMs during the projected period, 2014 - 2099, so may contain decimals.
}
\keyword{datasets}
