% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmoba}
\alias{fmoba}
\title{Annual burn area.}
\format{A data frame with 3,864 rows and 8 columns.}
\usage{
fmoba
}
\description{
A data frame of empirical historical annual burn area and ALFRESCO model burn area outputs for historical and projected periods, by FMO zone.
ID columns include ALFRESCO model set, FMO treatment, RCP, FMO zone, year, and the statistics: burn area (BA), cumulative burn area (CBA) and a 10-year moving average of inter-annual standard deviation in BA (BA_sd_ma10).
}
\keyword{datasets}
