% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmobaSummary}
\alias{fmobaSummary}
\title{Burn area summary statistics.}
\format{A data frame with 48 rows and 10 columns.}
\usage{
fmobaSummary
}
\description{
A data frame of burn area summary statistics by FMO zone. This data set is provided as a convenient summary of the \code{fmoba} data set.
It summarizes \code{fmoba} over time, providing the following statistics: minimum, first quartile, median, mean, third quartile and maximum burn area.
}
\keyword{datasets}
