% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{jfsp_plot}
\alias{jfsp_plot}
\title{Create plots based on various JFSP data sets}
\usage{
jfsp_plot(x, years = NULL, type = NULL, by_rcp = TRUE, col = NULL,
  file = NULL, base_size = 14, text_size = 18, ...)
}
\arguments{
\item{x}{a package data set. See details.}

\item{years}{numeric, vector of consecutive years. The maximum range is \code{1950:2099}. See details.}

\item{type}{character, the type of plot to make for a specific data set \code{x}. See details.}

\item{by_rcp}{logical, condition on RCP, defaults to \code{TRUE}. Otherwise marginalize over RCP. This applied to RCP-driven years, 2014 - 2099.}

\item{col}{optional vector of colors to override the defaults built into \code{jfsp_plot}.}

\item{file}{character, if provided, the plot is saved to disk. Otherwise, it is plotted in the R session graphics device. See details.}

\item{base_size}{base size passed to theme. See details.}

\item{text_size}{text size passed to theme.}

\item{...}{additional arguments. See details.}
}
\value{
a ggplot object. If saving a png file to disk, nothing is returned.
}
\description{
Create plots for each of the data sets included in the package using one convenient wrapper function.
}
\details{
Data sets included in the package can be plotted however desired, but for convenience the package offers this wrapper function.
When a package data set is paired with an appropriate plot \code{type},
a plot is created using standardized formatting based on the main \code{snapplot} package plot theme.
As long as the data set has not been manipulated, its assigned attributes will inform \code{jfsp_plot} how to graph it.

Data sets from the package and the associated available plot \code{type} options are as follows: ...

Additional arguments can be provided. General arguments include \code{family} (font family).
Arguments related to plots of a specific data set will be ignored for other data sets.
These include \code{breaks}, a vector of breaks applicable to time series plots with years along the x-axis,
\code{log = TRUE}, which can be used to select a log scale presentation for the burn area box plots type,
and \code{fmo}, which allows for subsetting the FMO zones available in the \code{fmoba} data set.
If not provided, it defaults to \code{fmo = c("Full", "Critical")} since these are the most important zones for work encapsulated by the package.

If the \code{showtext} is loaded, it may be necessary to significantly increase \code{base_size} and/or \code{text_size} for the 300 dpi image saved when \code{file} is not \code{NULL}.
Google fonts and even base R fonts may shrink significantly relative to the plot under these circumstances with showtext in effect (e.g., after calling \code{showtext_auto}).
If this happens, try 20 and 60 instead of 14 and 18, respectively.
This will be done for you automatically if you elect to name your font with the special name, \code{"gfont"}, and pass it as \code{family = "gfont"}.

When saving a plot directly from \code{jfsp_plot} by passing a filename to \code{file},
\code{jfsp_plot} will internally adjust plot formatting settings in order to save a 300 dpi resolution image to disk while maintaining consistent and appropriate sizing of text and other plot elements.
}
\examples{
jfsp_plot(fmoba, 1950:2013, "ba_box", log = TRUE)
}
